+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/meal/ui/MealScreen.kt
@@ -82,7 +82,9 @@ import com.maksimowiczm.foodyou.feature.addfood.SearchSharedTransition
 import com.maksimowiczm.foodyou.feature.addfood.ui.LocalAddFoodSharedTransitionScope
 import com.maksimowiczm.foodyou.feature.meal.domain.MealFood
 import com.maksimowiczm.foodyou.feature.meal.domain.MealWithFood
-import com.maksimowiczm.foodyou.feature.meal.ui.MealCardTransitionSpecs.overlayClipFromScreenToCard
+import com.maksimowiczm.foodyou.feature.meal.ui.card.MealCardTransitionKeys
+import com.maksimowiczm.foodyou.feature.meal.ui.card.MealCardTransitionSpecs
+import com.maksimowiczm.foodyou.feature.meal.ui.card.MealCardTransitionSpecs.overlayClipFromScreenToCard
 import com.maksimowiczm.foodyou.feature.meal.ui.component.MealHeader
 import com.maksimowiczm.foodyou.feature.meal.ui.component.NutrientsLayout
 import foodyou.app.generated.resources.*
